// src/data/enum.tsx
export enum InfoType {
  TITLE = "about-title",
  TAB = "about-tab",
  CONTACT = "contact",
  BANNER = "image-banner",
  LOGO = "image-logo",
  MEMBER="member",
}

export enum InfoAction {    
  SetAboutTitle = "SET_ABOUT_TITLE",
  SetAboutTab = "SET_ABOUT_TAB",
  SetContact = "SET_CONTACT",
  SetBanner = "SET_BANNER",
  SetLogo = "SET_LOGO", 
  SetMember="SET_MEMBER" 
}